Use single pip command to install packages
The existing method was using a single pip command for every package.
This could lead to errors, when the user try to fix a library A to a certain
version and another library B is also used, which referencing library B
(so this will use the latest allowed version of library A).
Now with the single call the pip command will use the correct versions.

Change-Id: If4abdcaf93b18bacff396222e1e7a991d96c7a1f